+++ packages/contracts/src/utils/verificationKey.ts
@@ -1,6 +1,6 @@
-import { FungibleToken, FungibleTokenAdmin, Pool, PoolFactory, PoolTokenHolder } from "../index.js"
+import { FungibleToken, FungibleTokenAdmin, Pool, PoolFactory, PoolTokenHolder } from "../indexpool.js"
 
-// node build/src/verificationKey.js
+// node dist/utils/verificationKey.js
 
 // get contract vk
 await PoolFactory.compile()
@@ -9,7 +9,7 @@ await FungibleToken.compile()
 await FungibleTokenAdmin.compile()
 const poolTokenHolderKey = await PoolTokenHolder.compile()
 
-console.log("pool key", poolKey)
+console.log("pool key", poolKey.verificationKey.data)
 console.log("pool key hash", poolKey.verificationKey.hash.toBigInt())
-console.log("pool token holder", poolTokenHolderKey)
+console.log("pool token holder", poolTokenHolderKey.verificationKey.data)
 console.log("pool token holder hash", poolTokenHolderKey.verificationKey.hash.toBigInt())